requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
aiohttp==3.8.5
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1